requirements.txt
numpy
torch
torchvision
tqdm
timm
pillow
pandas
opencv-python
matplotlib
torchmetrics
albumentations